# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KonbiniCompare API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface Category {
  /** Category identifier (e.g. "drinks", "snacks") */
  id: string;
  /** Display name in English */
  name: string;
  /** Display name in Japanese */
  nameJa: string;
  /** Emoji icon */
  icon: string;
  /** Which scoring dimensions apply to this category */
  scoringDimensions: string[];
}

export interface CategoriesResponse {
  categories: Category[];
}

export interface ProductNutrition {
  calories?: number | null;
  protein?: number | null;
  fat?: number | null;
  carbs?: number | null;
  sugar?: number | null;
  sodium?: number | null;
  caffeine?: number | null;
  fiber?: number | null;
}

export interface ProductSkincare {
  spf?: number | null;
  fragranceFree?: boolean | null;
  alcoholFree?: boolean | null;
  /** low | medium | high */
  irritationRisk?: string | null;
  keyIngredients?: string[];
}

export interface Product {
  id: string;
  name: string;
  nameJa?: string | null;
  brand?: string | null;
  category: string;
  /** 7-Eleven | FamilyMart | Lawson | Supermarket | Drugstore | etc. */
  store?: string | null;
  priceJpy?: number | null;
  volumeMl?: number | null;
  weightG?: number | null;
  servings?: number | null;
  nutrition?: ProductNutrition | null;
  skincare?: ProductSkincare | null;
  allergens: string[];
  additives: string[];
  imageUrl?: string | null;
  tags: string[];
}

export interface DemoProductsResponse {
  products: Product[];
  total: number;
}

export interface UserPreferences {
  /** 0.0 to 1.0 — how important is price */
  budgetSensitivity?: number;
  /** 0.0 to 1.0 — how important is nutritional quality */
  nutritionPriority?: number;
  /** 0.0 to 1.0 — how sensitive to caffeine (higher = prefer less caffeine) */
  caffeineSensitivity?: number;
  /** Allergens to flag (e.g. gluten, dairy, nuts) */
  allergenConcerns?: string[];
  /** 0.0 to 1.0 — prefer fewer/simpler additives */
  ingredientSimplicity?: number;
  /** 0.0 to 1.0 — how important is volume/weight per yen */
  valueForMoney?: number;
  /** 0.0 to 1.0 — prefer portable/single-serve options */
  convenience?: number;
  /** 0.0 to 1.0 — prefer low-irritation skincare (skincare category only) */
  skinSensitivity?: number;
  /** 0.0 to 1.0 — how important is protein content */
  proteinPriority?: number;
  /** 0.0 to 1.0 — prefer low-sugar options */
  sugarAvoidance?: number;
}

export interface ScoreBreakdown {
  price?: number | null;
  valueForMoney?: number | null;
  caffeine?: number | null;
  calories?: number | null;
  sugar?: number | null;
  protein?: number | null;
  additives?: number | null;
  allergenSafety?: number | null;
  skinSafety?: number | null;
  convenience?: number | null;
  missingDataPenalty?: number | null;
}

export interface RankedResult {
  rank: number;
  productId: string;
  productName: string;
  /** 0.0 to 1.0 weighted total score */
  score: number;
  scoreBreakdown: ScoreBreakdown;
  strengths: string[];
  tradeoffs: string[];
  cautions: string[];
  missingFields: string[];
  isBestMatch: boolean;
}

export interface CompareProductsBody {
  category: string;
  /**
   * IDs of products from demo catalog to compare
   * @minItems 2
   * @maxItems 5
   */
  productIds: string[];
  preferences?: UserPreferences;
}

/**
 * Field-by-field comparison matrix
 */
export type CompareProductsResponseComparisonTable = { [key: string]: unknown };

export interface CompareProductsResponse {
  category: string;
  products: Product[];
  rankedResults: RankedResult[];
  /** Product name of the top-ranked product */
  bestMatch: string;
  /** Japanese summary explanation */
  summaryJa: string;
  /** English summary explanation */
  summaryEn: string;
  /** Field-by-field comparison matrix */
  comparisonTable: CompareProductsResponseComparisonTable;
  /** mock | live | fallback */
  processingMode: string;
  preferences?: UserPreferences;
}

export interface DemoScenario {
  id: string;
  title: string;
  titleJa: string;
  description: string;
  category: string;
  productIds: string[];
  preferences?: UserPreferences;
}

export interface DemoScenariosResponse {
  scenarios: DemoScenario[];
}

export type GetDemoProductsParams = {
  category?: string;
};